+++ quint/src/effects/NondetChecker.ts
@@ -5,7 +5,7 @@
  * --------------------------------------------------------------------------------- */
 
 /**
- * Checking for the misuse of 'nondet', 'oneOf', and 'generate'.
+ * Checking for the misuse of 'nondet', 'oneOf', and 'apalache::generate'.
  * Necessary to ensure they are compatible with the exists operator from TLA+.
  *
  * @author Gabriela Moreira
@@ -32,7 +32,7 @@ export class NondetChecker implements IRVisitor {
   }
 
   /**
-   * Checks declarations for misuse of 'nondet', 'oneOf', and 'generate'.
+   * Checks declarations for misuse of 'nondet', 'oneOf', and 'apalache::generate'.
    *
    * @param types - the types of the declarations
    * @param declarations - the declarations to check
@@ -66,7 +66,7 @@ export class NondetChecker implements IRVisitor {
   }
 
   enterApp(app: QuintApp) {
-    if (app.opcode !== 'oneOf' && app.opcode !== 'generate') {
+    if (app.opcode !== 'oneOf' && app.opcode !== 'apalache::generate') {
       // nothing to check
       return
     }
@@ -87,12 +87,12 @@ export class NondetChecker implements IRVisitor {
       return
     }
 
-    // body of nondet must be an application of oneOf or generate
+    // body of nondet must be an application of oneOf or apalache::generate
     const body = expr.opdef.expr
-    if (body.kind !== 'app' || (body.opcode !== 'oneOf' && body.opcode !== 'generate')) {
+    if (body.kind !== 'app' || (body.opcode !== 'oneOf' && body.opcode !== 'apalache::generate')) {
       this.errors.push({
         code: 'QNT204',
-        message: `the outermost expression in a nondet definition must be either 'oneOf' or 'generate'`,
+        message: `the outermost expression in a nondet definition must be either 'oneOf' or 'apalache::generate'`,
         reference: body.id,
         data: {},
       })